recurrent quant int8->bf16

+++ vllm_mlx/turn_prefix_cache.py
@@ -64,7 +64,7 @@ class TurnPrefixCacheConfig:
     checkpoint_stride: int = 4000 # tokens between permanent checkpoints; 0 = every node
     max_memory_gb: float = 8.0
     kv_dtype: str = "int8"                      # "bf16" or "int8"
-    recurrent_dtype: str = "int8"                # "none", "fp16", "bf16", or "int8" (per-channel)
+    recurrent_dtype: str = "bf16"                # "none", "fp16", "bf16", or "int8" (per-channel)
     persist_dir: str | None = None    # None = disabled
     ssd_max_gb: float = 0.0           # 0 = disabled
     ssd_dir: str | None = None        # SSD spill directory (defaults to persist_dir/ssd)
